Write a program that reads a list of integers and two numbers: p and m. The program is to check that p and m never occur next to each other (in any order) in the list.
Input data format

The first line contains the size of a list, N.
The next N lines contain the elements of the list.
The last line contains two integer numbers p and m separated by space.

The result is a single value, that is, "YES" if p and m never occur next to each other, or "NO" if otherwise.